move buffer pool logic into own pkg
